Add filter metadata and add Ubiquiti tracking domain
The filter did not look great when adding it to your add blocker of choice, because metadata was absent. The code contained in this commit ensures that ad blockers can now properly identify what this list is for, where it originates from and when it was last updated. It also adds a rule for blocking Ubiquiti their tracking domain (see https://twitter.com/unserkaiser/status/1366672287156551681)

* Add filter metadata
* Tell Ubiquiti that GDPR violations are not cool, yo